+++ Sample/View Controllers/Table View/DeepDiffTableProviderSampleViewController.swift
@@ -105,7 +105,7 @@ final class DeepDiffTableProviderSampleViewController: UIViewController {
             return cell
         }
         
-        tableProvider.configureSectionHeader = { provider, section in
+        tableProvider.configureSectionHeader = { _, _, _, section in
             let container = UIView()
             let label = UILabel()
             label.translatesAutoresizingMaskIntoConstraints = false
@@ -117,7 +117,7 @@ final class DeepDiffTableProviderSampleViewController: UIViewController {
             label.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: -15).isActive = true
             label.rightAnchor.constraint(equalTo: container.rightAnchor, constant: -15).isActive = true
 
-            label.text = provider.sections[section].title
+            label.text = section.title
 
             return container
         }