+++ Sample/View Controllers/Table View/DiffableTableViewSampleController.swift
@@ -96,7 +96,7 @@ final class DiffableTableViewSampleController: UIViewController {
             return cell
         }
         
-        tableProvider.configureSectionHeader = { provider, section in
+        tableProvider.configureSectionHeader = { _, _, _, section in
             let container = UIView()
             let label = UILabel()
             label.translatesAutoresizingMaskIntoConstraints = false
@@ -108,7 +108,7 @@ final class DiffableTableViewSampleController: UIViewController {
             label.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: -15).isActive = true
             label.rightAnchor.constraint(equalTo: container.rightAnchor, constant: -15).isActive = true
 
-            label.text = provider.sections[section].title
+            label.text = section.title
 
             return container
         }