+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.4] 07.10.2020
+### Added
+-  add `TableView`,  `SectionIndex`,  `Section` to `configureSectionHeader` and `configureSectionFooterHeight`
+
+## [0.0.3] 12.8.2020
+### Added
+-  add `IndexPath` to `configureOnItemSelected` 
+
 ## [0.0.2] 31.05.2020
 
 ### Changed